backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
black==22.3.0
click==8.1.3
Django==4.0.5
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
simplejwt==2.0.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.2.0